fleshing out systems and resources

+++ systems/system.go
@@ -3,5 +3,5 @@ package systems
 import "BasicECS/entities"
 
 type System interface {
-	Update(dt float32, entityManager *entities.EntityManager)
+	Update(dt float64, entityManager *entities.EntityManager)
 }
